requirements/dev.txt
-r common.txt
ForgeryPy==0.1
Pygments==1.6
colorama==0.2.7
coverage==3.7.1
httpie==0.7.2
requests==2.1.0
selenium==2.45.0
flask-api==0.6.9
requests==2.13.0